Module1_IntroToMLOps/ocr-mlops-project/requirements.txt
pytesseract
Pillow
opencv-python-headless
numpy
pytest
black
pylint
